requirements.txt
flask
requests
Pillow
face-recognition
face_recognition_models
# git+https://github.com/ageitgey/face_recognition_models
psycopg2